+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/controller/Oauth2Controller.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @RestController
@@ -16,8 +17,10 @@ import java.io.IOException;
 public class Oauth2Controller {
     private final Oauth2NotWebService oauth2NotWebService;
     @GetMapping("/auth/{provider}") //GOOGLE이 들어올 것이다.
-    public Response socialLoginRedirect(@PathVariable String provider , @RequestParam String code ) throws IOException {
+    public Response socialLoginRedirect(@PathVariable String provider , @RequestParam String code,
+    HttpServletResponse response) throws IOException {
         ProviderType providerType = ProviderType.valueOf(provider.toUpperCase());
-        return Response.success("data", oauth2NotWebService.signInByOAuth(providerType, code));
+        oauth2NotWebService.signInByOAuth(providerType, code , response);
+        return Response.success("message", "SUCCESS" );
     }
 }
