+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/handler/Oauth2Handler.java
@@ -2,9 +2,8 @@ package com.dju.gdsc.domain.oauth.handler;
 
 import com.dju.gdsc.domain.oauth.entity.ProviderType;
 import com.dju.gdsc.domain.oauth.info.OAuth2UserInfo;
-import com.dju.gdsc.domain.oauth.infrastructure.OAuthRequester;
+import com.dju.gdsc.domain.oauth.infrastructure.OauthRequester;
 import lombok.RequiredArgsConstructor;
-import okhttp3.internal.http2.ErrorCode;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -12,14 +11,14 @@ import java.util.List;
 @Component
 @RequiredArgsConstructor
 public class Oauth2Handler {
-    private final List<OAuthRequester> oAuthRequesters;
+    private final List<OauthRequester> oAuthRequesters;
 
     public OAuth2UserInfo getUserInfo(final ProviderType provider, final String code) {
-        OAuthRequester requester = getRequester(provider);
+        OauthRequester requester = getRequester(provider);
         return requester.getUserInfo(code);
     }
 
-    private OAuthRequester getRequester(final ProviderType provider) {
+    private OauthRequester getRequester(final ProviderType provider) {
         return oAuthRequesters.stream()
                 .filter(requester -> requester.supports(provider))
                 .findFirst()
