+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/infrastructure/OauthRequester.java
@@ -3,7 +3,7 @@ package com.dju.gdsc.domain.oauth.infrastructure;
 import com.dju.gdsc.domain.oauth.entity.ProviderType;
 import com.dju.gdsc.domain.oauth.info.OAuth2UserInfo;
 
-public interface OAuthRequester {
+public interface OauthRequester {
     boolean supports(ProviderType provider);
 
     OAuth2UserInfo getUserInfo(String code);
