+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/service/Oauth2NotWebService.java
@@ -1,31 +1,21 @@
 package com.dju.gdsc.domain.oauth.service;
 
-import com.dju.gdsc.domain.common.dto.Response;
 import com.dju.gdsc.domain.common.properties.AppProperties;
 import com.dju.gdsc.domain.member.entity.Member;
-import com.dju.gdsc.domain.member.entity.MemberInfo;
-import com.dju.gdsc.domain.member.model.RoleType;
 import com.dju.gdsc.domain.member.repository.MemberRepository;
-import com.dju.gdsc.domain.member.service.MemberService;
 import com.dju.gdsc.domain.oauth.dto.TokenResponseDto;
 import com.dju.gdsc.domain.oauth.entity.ProviderType;
 import com.dju.gdsc.domain.oauth.entity.UserRefreshToken;
 import com.dju.gdsc.domain.oauth.handler.Oauth2Handler;
 import com.dju.gdsc.domain.oauth.info.OAuth2UserInfo;
-import com.dju.gdsc.domain.oauth.info.OAuth2UserInfoFactory;
-import com.dju.gdsc.domain.oauth.infrastructure.GoogleOAuthRequester;
 import com.dju.gdsc.domain.oauth.repository.UserRefreshTokenRepository;
 import com.dju.gdsc.domain.oauth.token.AuthToken;
 import com.dju.gdsc.domain.oauth.token.AuthTokenProvider;
-import com.nimbusds.oauth2.sdk.TokenResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.time.LocalDateTime;
 import java.util.Date;
 
 @Service
@@ -39,11 +29,15 @@ public class Oauth2NotWebService {
     private final AuthTokenProvider tokenProvider;
     private final UserRefreshTokenRepository userRefreshTokenRepository;
     @Transactional
-    public TokenResponseDto signInByOAuth(final ProviderType provider, final String code) {
+    public void signInByOAuth(final ProviderType provider, final String code , HttpServletResponse response)  {
         OAuth2UserInfo userInfo = oauth2Handler.getUserInfo(provider, code);
         Member member = memberRepository.findByProviderTypeAndUserId(provider, userInfo.getId())
                 .orElseGet(() -> customOAuth2UserService.createUser(userInfo , provider));
-        return createNewTokens(member);
+        TokenResponseDto token = createNewTokens(member);
+        response.addHeader("token", token.getAccessToken());
+        response.addHeader("refreshToken", token.getRefreshToken());
+        response.addHeader("expiresIn", String.valueOf(token.getExpireIn()));
+
     }
 
     private TokenResponseDto createNewTokens(final Member member) {
